Move project items dialog to its own file.
